Switched money stuff back from decimal to float, to be better supported by Unity's serialization.

+++ Assets/Therapy Receipt Tycoon/BudgetScreen.cs
@@ -6,15 +6,15 @@ public class BudgetScreen : MonoBehaviour {
 	/// <summary>
 	/// The amount of funds that were available at the start of the round, used for stat tracking.
 	/// </summary>
-	public decimal m_fundsAtStartofRound = 1000;
+	public float m_fundsAtStartofRound = 1000;
 	/// <summary>
 	/// Currently availableFunds
 	/// </summary>
-	public decimal availableFunds = 1000;
+	public float availableFunds = 1000;
 	/// <summary>
 	/// The cost of running a therapy business. Should be set from round settings.
 	/// </summary>
-	public decimal currentCost;
+	public float currentCost;
 	
 	public bool IsGameOver{
 		get { return availableFunds < 0; }
