+++ Assets/Therapy Receipt Tycoon/Rounds/GameRound.cs
@@ -6,10 +6,10 @@ public class GameRound : ScriptableObject {
 
 	public int clientsPerRound = 8;
 	
-	public decimal regularPrice = 200;
-	public decimal higherPrice = 300;
-	public decimal lowerPrice = 150;
-	public decimal businessCosts = 600;
+	public float regularPrice = 200;
+	public float higherPrice = 300;
+	public float lowerPrice = 150;
+	public float businessCosts = 600;
 	
 	public float highPriceInsanityAdjustment = 0.1f;
 	public float lowPriceInsanityAdjustment = -0.1f;
